* ad_print.c (ad_print): Cast arguments to printf() to longs to 	match format string.
git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@13310 dc483132-0cff-0310-8789-dd5450dbe970

+++ src/lib/krb4/ad_print.c
@@ -34,7 +34,7 @@ ad_print(x)
     printf("\n%s %s %s ", x->pname, x->pinst, x->prealm);
     far_fputs (inet_ntoa(ina), stdout);
     printf(" flags %u cksum 0x%lX\n\ttkt_tm 0x%lX sess_key",
-           x->k_flags, x->checksum, x->time_sec);
+            x->k_flags, (long) x->checksum, (long) x->time_sec);
     printf("[8] =");
 #ifdef NOENCRYPTION
     placebo_cblock_print(x->session);
